+++ nflow-engine/src/main/resources/scripts/db/db2.create.ddl.sql
@@ -17,6 +17,7 @@ create table nflow_workflow (
   retries int not null default 0,
   created timestamp(3) not null default current_timestamp,
   modified timestamp(3) not null default current_timestamp,
+  started timestamp(3),
   executor_group varchar(64) not null,
   workflow_signal int,
   constraint nflow_workflow_uniq unique (type, external_id, executor_group)
@@ -112,6 +113,7 @@ create table nflow_archive_workflow (
   retries int not null default 0,
   created timestamp(3) not null,
   modified timestamp(3) not null,
+  started timestamp(3),
   executor_group varchar(64) not null,
   workflow_signal int,
   constraint nflow_archive_workflow_uniq unique (type, external_id, executor_group)