+++ nflow-engine/src/main/resources/scripts/db/mysql.legacy.create.ddl.sql
@@ -16,6 +16,7 @@ create table if not exists nflow_workflow (
   retries int not null default 0,
   created timestamp not null,
   modified timestamp not null default current_timestamp on update current_timestamp,
+  started timestamp,
   executor_group varchar(64) not null,
   workflow_signal int,
   constraint nflow_workflow_uniq unique (type, external_id, executor_group)
@@ -108,6 +109,7 @@ create table if not exists nflow_archive_workflow (
   retries int not null default 0,
   created timestamp not null,
   modified timestamp not null,
+  started timestamp,
   executor_group varchar(64) not null,
   workflow_signal int,
   constraint nflow_archive_workflow_uniq unique (type, external_id, executor_group)