+++ nflow-engine/src/main/resources/scripts/db/update-5.5.0-x/mysql.update.ddl.sql
@@ -1,0 +1,1 @@
+alter table nflow_workflow add started timestamp(3);