+++ nflow-engine/src/main/resources/scripts/db/update-5.5.0-x/sqlserver.update.ddl.sql
@@ -1,0 +1,1 @@
+alter table nflow_workflow add started datetimeoffset(3);